Ensure that prematurely spawned actors are also deferred. Fixes #1281
