+++ packages/xstate-react/test/useService.test.tsx
@@ -182,101 +182,4 @@ describe('useService hook', () => {
     fireEvent.click(incButton);
     expect(countEl.textContent).toBe('1');
   });
-
-  it('initial invoked actor should be immediately available', (done) => {
-    const childMachine = createMachine({
-      id: 'childMachine',
-      initial: 'active',
-      states: {
-        active: {}
-      }
-    });
-    const machine = createMachine({
-      initial: 'active',
-      invoke: {
-        id: 'child',
-        src: childMachine
-      },
-      states: {
-        active: {}
-      }
-    });
-
-    const ChildTest: React.FC<{ actor: Actor<any> }> = ({ actor }) => {
-      const [state] = useActor(actor);
-
-      expect(state.value).toEqual('active');
-
-      done();
-
-      return null;
-    };
-
-    const Test = () => {
-      const [state] = useMachine(machine);
-
-      return <ChildTest actor={state.children.child} />;
-    };
-
-    render(
-      <React.StrictMode>
-        <Test />
-      </React.StrictMode>
-    );
-  });
-
-  it('invoked actor should be able to receive (deferred) events that it replays when active', (done) => {
-    const childMachine = createMachine({
-      id: 'childMachine',
-      initial: 'active',
-      states: {
-        active: {
-          on: {
-            FINISH: { actions: sendParent('FINISH') }
-          }
-        }
-      }
-    });
-    const machine = createMachine({
-      initial: 'active',
-      invoke: {
-        id: 'child',
-        src: childMachine
-      },
-      states: {
-        active: {
-          on: { FINISH: 'success' }
-        },
-        success: {}
-      }
-    });
-
-    const ChildTest: React.FC<{ actor: Actor<any> }> = ({ actor }) => {
-      const [state, send] = useActor(actor);
-
-      expect(state.value).toEqual('active');
-
-      React.useEffect(() => {
-        send({ type: 'FINISH' });
-      }, []);
-
-      return null;
-    };
-
-    const Test = () => {
-      const [state] = useMachine(machine);
-
-      if (state.matches('success')) {
-        done();
-      }
-
-      return <ChildTest actor={state.children.child} />;
-    };
-
-    render(
-      <React.StrictMode>
-        <Test />
-      </React.StrictMode>
-    );
-  });
 });
